Be clearer about LTO
The LTO section currently makes it seem like `lto = true` is the only type of LTO and equivalent to the LTO that's linked on the LLVM page.
This is not true. On the rustc side, we have "thin" LTO and "fat" LTO, (`true` meaning fat). Neither of these actually involve the linker, they happen entirely on the rustc and LLVM side.
Fat LTO is what is best for binary size, while thin is great for the performance of larger programs.
Neither of the two is linker plugin LTO (what's meant on the LLVM page), which is enabled with ,`-Clinker-plugin-lto` and actually involves the linker calling back into LLVM, which enables cross-language LTO.

I made the suggested config `"fat"` because I think `true` is bad as it is not explicit that the user knows what they really want.